requirements.txt
flask
langchain
python-dotenv
pymilvus
streamlit
python-dotenv